+++ src/main/java/superstartrek/client/model/StarMap.java
@@ -9,8 +9,7 @@ public class StarMap {
 
 	protected Quadrant[][] quadrants = new Quadrant[Constants.SECTORS_EDGE][Constants.SECTORS_EDGE];
 	public Enterprise enterprise;
-	//TODO: this should probably be an int
-	protected long starDate = 2100;
+	protected int starDate = 2100;
 
 	public static boolean within_distance(int x1, int y1, int x2, int y2, double range) {
 		int dx = x1 - x2;
@@ -52,15 +51,15 @@ public class StarMap {
 		quadrants[q.getX()][q.getY()] = q;
 	}
 
-	public long getStarDate() {
+	public int getStarDate() {
 		return starDate;
 	}
 	
-	public void setStarDate(long sd) {
+	public void setStarDate(int sd) {
 		this.starDate = sd;
 	}
 
-	public void advanceStarDate(long value) {
+	public void advanceStarDate(int value) {
 		starDate += value;
 	}
